+++ UntitledTemalab/app/src/main/java/hu/bme/aut/untitledtemalab/network/FreelancerAPI.kt
@@ -11,29 +11,29 @@ interface FreelancerAPI {
 
     @GET("/user/{user-id}/jobs/sent")
     fun getUsersSentJobs(
-        @Path("user-id") userId: Int
+        @Path("user-id") userId: Long
     ): Call<MutableList<JobData>>
 
     @GET("/user/{user-id}/jobs/delivered")
     fun getUsersDeliveredJobs(
-        @Path("user-id") userId: Int
+        @Path("user-id") userId: Long
     ): Call<MutableList<JobData>>
 
     @GET("/user/{user-id}/jobs/sent")
     fun getUsersPostedJobsByStatus(
-        @Path("user-id") userId: Int,
+        @Path("user-id") userId: Long,
         @Query("status") status: String
     ): Call<MutableList<JobData>>
 
     @GET("/user/{user-id}/jobs/delivered")
     fun getUsersAcceptedJobsByStatus(
-        @Path("user-id") userId: Int,
+        @Path("user-id") userId: Long,
         @Query("status") status: String
     ): Call<MutableList<JobData>>
 
     @GET("/user/profile")
     fun getUserProfileById(
-        @Query("user-id") userId: Int
+        @Query("user-id") userId: Long
     ): Call<UserData>
 
     @GET("/jobs")
